Refactor how modals are rendered in order to support keyboard shortcuts
