+++ app/javascript/components/dropdowns/track-menu/LeaveTrackModal.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { Modal, ModalProps } from '../../modals/Modal'
 import { useMutation } from 'react-query'
 import { useIsMounted } from 'use-is-mounted'
@@ -7,36 +7,15 @@ import { typecheck } from '../../../utils/typecheck'
 import { FormButton } from '../../common/FormButton'
 import { ErrorBoundary, ErrorMessage } from '../../ErrorBoundary'
 
-export const ResetTrackButton = ({
-  endpoint,
-}: {
-  endpoint: string
-}): JSX.Element => {
-  const [open, setOpen] = useState(false)
-
-  return (
-    <React.Fragment>
-      <button type="button" onClick={() => setOpen(true)}>
-        Reset track
-      </button>
-      <ResetTrackModal
-        open={open}
-        onClose={() => setOpen(false)}
-        endpoint={endpoint}
-      />
-    </React.Fragment>
-  )
-}
-
 type UserTrack = {
   links: {
     self: string
   }
 }
 
-const DEFAULT_ERROR = new Error('Unable to reset track')
+const DEFAULT_ERROR = new Error('Unable to leave track')
 
-export const ResetTrackModal = ({
+export const LeaveTrackModal = ({
   endpoint,
   ...props
 }: Omit<ModalProps, 'className'> & { endpoint: string }): JSX.Element => {
@@ -68,9 +47,9 @@ export const ResetTrackModal = ({
   )
 
   return (
-    <Modal className="m-reset-track" {...props}>
+    <Modal className="m-leave-track" {...props}>
       <FormButton onClick={() => mutation()} status={status}>
-        Reset track
+        Leave track
       </FormButton>
       <ErrorBoundary>
         <ErrorMessage error={error} defaultError={DEFAULT_ERROR} />
